api/dashboard/requirements.txt
Flask==2.3.2
pymongo==4.6.0
flask-login==0.6.2
paho-mqtt==1.6.1
